+++ src/lib.rs
@@ -1,11 +1,8 @@
 //! A DNS library for Rust.
 //!
-//! This crates provides a number of building blocks for developing
+//! This crate provides a number of building blocks for developing
 //! functionality related to the
 //! [Domain Name System (DNS)](https://www.rfc-editor.org/rfc/rfc9499.html).
-//! It provides fundamental types, traits, and code as well as a wide range
-//! of optional features. The intent is to eventually cover all aspects of
-//! modern DNS.
 //!
 //! The crate uses feature flags to allow you to select only those modules
 //! you need for you particular project. In most cases, the feature names
@@ -25,7 +22,8 @@
 //! more specific features that are not required in all applications. In order
 //! to keep the amount of code to be compiled and the number of dependencies
 //! small, these are hidden behind feature flags through which they can be
-//! enabled if required. The flags have the same names as the modules.
+//! enabled if required. The flags have the same names as the modules or the
+//! name prefixed with 'unstable-' if the module is still under development.
 //!
 //! Currently, there are the following modules:
 //!
@@ -36,18 +34,14 @@
 #![cfg_attr(not(feature = "resolv"), doc = "* resolv:")]
 //!   An asynchronous DNS resolver based on the
 //!   [Tokio](https://tokio.rs/) async runtime.
-#![cfg_attr(feature = "sign", doc = "* [sign]:")]
-#![cfg_attr(not(feature = "sign"), doc = "* sign:")]
-//!   Experimental support for DNSSEC signing.
+#![cfg_attr(feature = "unstable-crypto", doc = "* [crypto]:")]
+#![cfg_attr(not(feature = "unstable-crypto"), doc = "* crypto:")]
+//!   Experimental support for cryptographic backends, key generation and
+//!   import.
+//! * [dnssec]: DNSSEC signing and validation.
 #![cfg_attr(feature = "tsig", doc = "* [tsig]:")]
 #![cfg_attr(not(feature = "tsig"), doc = "* tsig:")]
 //!   Support for securing DNS transactions with TSIG records.
-#![cfg_attr(feature = "validate", doc = "* [validate]:")]
-#![cfg_attr(not(feature = "validate"), doc = "* validate:")]
-//!   Experimental support for DNSSEC validation.
-#![cfg_attr(feature = "unstable-validator", doc = "* [validator]:")]
-#![cfg_attr(not(feature = "unstable-validator"), doc = "* validator:")]
-//!   A DNSSEC validator.
 #![cfg_attr(feature = "zonefile", doc = "* [zonefile]:")]
 #![cfg_attr(not(feature = "zonefile"), doc = "* zonefile:")]
 //!   Experimental reading and writing of zone files, i.e. the textual
@@ -61,61 +55,79 @@
 //!
 //! # Reference of feature flags
 //!
-//! The following is the complete list of the feature flags with the
-//! exception of unstable features which are described below.
+//! Several feature flags simply enable support for other crates, e.g. by
+//! adding `impl`s for their types.  They are optional and do not introduce
+//! new functionality into this crate.
 //!
 //! * `bytes`: Enables using the types `Bytes` and `BytesMut` from the
 //!   [bytes](https://github.com/tokio-rs/bytes) crate as octet sequences.
-//! * `chrono`: Adds the [chrono](https://github.com/chronotope/chrono)
-//!   crate as a dependency. This adds support for generating serial numbers
-//!   from time stamps.
+//!
 //! * `heapless`: enables the use of the `Vec` type from the
 //!   [heapless](https://github.com/japaric/heapless) crate as octet
 //!   sequences.
-//! * `interop`: Activate interoperability tests that rely on other software
-//!   to be installed in the system (currently NSD and dig) and will fail if
-//!   it isn’t. This feature is not meaningful for users of the crate.
+//!
+//! * `smallvec`: enables the use of the `Smallvec` type from the
+//!   [smallvec](https://github.com/servo/rust-smallvec) crate as octet
+//!   sequences.
+//!
+//! Some flags enable support for specific kinds of operations that are not
+//! otherwise possible.  They are gated as they may not always be necessary
+//! and they may introduce new dependencies.
+//!
+//! * `chrono`: Adds the [chrono](https://github.com/chronotope/chrono)
+//!   crate as a dependency. This adds support for generating serial numbers
+//!   from time stamps.
+//!
 //! * `rand`: Enables a number of methods that rely on a random number
 //!   generator being available in the system.
-//! * `resolv`: Enables the asynchronous stub resolver via the
-#![cfg_attr(feature = "resolv", doc = "  [resolv]")]
-#![cfg_attr(not(feature = "resolv"), doc = "  resolv")]
-//!   module.
-//! * `resolv-sync`: Enables the synchronous version of the stub resolver.
-//! * `ring`: Enables crypto functionality via the
-//!   [ring](https://github.com/briansmith/ring) crate.
+//!
 //! * `serde`: Enables serde serialization for a number of basic types.
-//! * `sign`: basic DNSSEC signing support. This will enable the
-#![cfg_attr(feature = "sign", doc = "  [sign]")]
-#![cfg_attr(not(feature = "sign"), doc = "  sign")]
-//!   module and requires the `std` feature. Note that this will not directly
-//!   enable actual signing. For that you will also need to pick a crypto
-//!   module via an additional feature. Currently we only support the `ring`
-//!   module, but support for OpenSSL is coming soon.
+//!
 //! * `siphasher`: enables the dependency on the
 //!   [siphasher](https://github.com/jedisct1/rust-siphash) crate which allows
 //!   generating and checking hashes in [standard server
 //!   cookies][crate::base::opt::cookie::StandardServerCookie].
-//! * `smallvec`: enables the use of the `Smallvec` type from the
-//!   [smallvec](https://github.com/servo/rust-smallvec) crate as octet
-//!   sequences.
+//!
 //! * `std`: support for the Rust std library. This feature is enabled by
 //!   default.
+//!
+//! A special case here is cryptographic backends.  Certain modules (e.g. for
+//! DNSSEC signing and validation) require a backend to provide cryptography.
+//! At least one such module should be enabled.
+//!
+//! * `openssl`: Enables crypto functionality via OpenSSL through the
+//!   [rust-openssl](https://github.com/sfackler/rust-openssl) crate.
+//!
+//! * `ring`: Enables crypto functionality via the
+//!   [ring](https://github.com/briansmith/ring) crate.
+//!
+//! Some flags represent entire categories of functionality within this crate.
+//! Each flag is associated with a particular module.  Note that some of these
+//! modules are under heavy development, and so have unstable feature flags
+//! which are categorized separately.
+//!
+//! * `net`: Enables sending and receiving DNS messages via the
+#![cfg_attr(feature = "net", doc = "  [net]")]
+#![cfg_attr(not(feature = "net"), doc = "  net")]
+//!   module.
+//!
+//! * `resolv`: Enables the asynchronous stub resolver via the
+#![cfg_attr(feature = "resolv", doc = "  [resolv]")]
+#![cfg_attr(not(feature = "resolv"), doc = "  resolv")]
+//!   module.
+//!
+//!   * `resolv-sync`: Enables the synchronous version of the stub resolver.
+//!
 //! * `tsig`: support for signing and validating message exchanges via TSIG
 //!   signatures. This enables the
 #![cfg_attr(feature = "tsig", doc = "  [tsig]")]
 #![cfg_attr(not(feature = "tsig"), doc = "  tsig")]
-//!   module and currently pulls in the
-//!   `bytes`, `ring`, and `smallvec` features.
-//! * `validate`: basic DNSSEC validation support. This feature enables the
-#![cfg_attr(feature = "validate", doc = "  [validate]")]
-#![cfg_attr(not(feature = "validate"), doc = "  validate")]
-//!   module and currently also enables the `std` and `ring`
-//!   features.
+//!   module and currently enables `bytes`, `ring`, and `smallvec`.
+//!
 //! * `zonefile`: reading and writing of zonefiles. This feature enables the
 #![cfg_attr(feature = "zonefile", doc = "  [zonefile]")]
 #![cfg_attr(not(feature = "zonefile"), doc = "  zonefile")]
-//!   module and currently also enables the `bytes` and `std` features.
+//!   module and currently also enables `bytes`, `serde`, and `std`.
 //!
 //! # Unstable features
 //!
@@ -137,6 +149,18 @@
 //!   a client perspective; primarily the `net::client` module.
 //! * `unstable-server-transport`: receiving and sending DNS messages from
 //!   a server perspective; primarily the `net::server` module.
+//! * `unstable-crypto`: this feature flag needs to be combined with one or
+//!   more feature flags that enable cryptographic backends (currently `ring`
+//!   and `openssl`). This feature flags enables all parts of the crypto
+//!   module except for private key generation and signing.
+//! * `unstable-crypto-sign`: this feature flag needs to be combined with one
+//!   or more feature flags that enable cryptographic backends. This feature
+//!   flag enables all parts of the crypto module.
+//! * `unstable-sign`: basic DNSSEC signing support. This will enable the
+//!   `dnssec::sign`
+//!   module and requires the `std` feature. In order to actually perform any
+//!   signing, also enable one or more cryptographic backend modules (`ring`
+//!   and `openssl`). Enabling this will also enable `unstable-crypto-sign`.
 //! * `unstable-validator`: a DNSSEC validator, primarily the `validator`
 //!   and the `net::client::validator` modules.
 //! * `unstable-xfr`: zone transfer related functionality..
@@ -163,15 +187,14 @@ extern crate std;
 extern crate core;
 
 pub mod base;
+pub mod crypto;
 pub mod dep;
+pub mod dnssec;
 pub mod net;
 pub mod rdata;
 pub mod resolv;
-pub mod sign;
 pub mod stelline;
 pub mod tsig;
 pub mod utils;
-pub mod validate;
-pub mod validator;
 pub mod zonefile;
 pub mod zonetree;