+++ Changelog.md
@@ -4,16 +4,40 @@
 
 Breaking changes
 
+* FIX: Use base 16 per RFC 4034 for the DS digest, not base 64. ([#423])
+* Stricter RFC 1035 compliance by default in the `Zonefile` parser. ([#477])
+
 New
 
 * Added `HashCompressor`, an unlimited name compressor that uses a hash map
   rather than a tree. ([#396])
 * Changed `fmt::Display` for `HINFO` records to a show a quoted string.
   ([#421])
+* Added support for `NAPTR` record type. ([#427] by [@weilence])
+* Added initial fuzz testing support for some types via a new `arbitrary`
+  feature (not enabled by default). ([#441])
+* Added `StubResolver::add_connection()` to allow adding a connection to the
+  running resolver. In combination with `ResolvConf::new()` this can also be
+  used to control the connections made when testing code that uses the stub
+  resolver. ([#440])
+* Add `ZonefileFmt` trait for printing records as zonefiles. ([#379], [#446],
+  [#463])
 
 Bug fixes
 
+* NSEC records should include themselves in the generated bitmap. ([#417])
+* Trailing double quote wrongly preserved when parsing record data. ([#470],
+  [#472])
+* Don't error with unexpected end of entry for RFC 3597 RDATA of length zero. ([475])
+
 Unstable features
+
+* New unstable feature `unstable-crypto` that enable cryptography support
+  for features that do not rely on secret keys. This feature needs either
+  or both of the features `ring` and `openssl` ([#416])
+* New unstable feature `unstable-crypto-sign` that enable cryptography support
+  including features that rely on secret keys. This feature needs either
+  or both of the features `ring` and `openssl` ([#416])
 
 * `unstable-server-transport`
   * The trait `SingleService` which is a simplified service trait for
@@ -27,11 +51,58 @@ Unstable features
   * A sample query router, called `QnameRouter`, that routes requests based
     on the QNAME field in the request ([#353]).
 
+* `unstable-client-transport`
+  * introduce timeout option in multi_stream ([#424]).
+  * improve probing in redundant ([#424]).
+  * restructure configuration for multi_stream and redundant ([#424]).
+  * introduce a load balancer client transport. This transport tries to
+    distribute requests equally over upstream transports ([#425]).
+
+* `unstable-sign`
+  * add key lifecycle management ([#459]).
+  * add support for adding NSEC3 records when signing.
+  * add support for ZONEMD.
+
+* `unstable-validator`
+  * The `validate` crate is moved to `dnssec::validator::base`.
+  * The `validator` crate is moved to `dnssec::validator`.
+
 Other changes
 
 [#353]: https://github.com/NLnetLabs/domain/pull/353
+[#379]: https://github.com/NLnetLabs/domain/pull/379
 [#396]: https://github.com/NLnetLabs/domain/pull/396
-[#421]: https://github.com/NLnetLabs/domain/pull/412
+[#416]: https://github.com/NLnetLabs/domain/pull/416
+[#417]: https://github.com/NLnetLabs/domain/pull/417
+[#421]: https://github.com/NLnetLabs/domain/pull/421
+[#423]: https://github.com/NLnetLabs/domain/pull/423
+[#424]: https://github.com/NLnetLabs/domain/pull/424
+[#425]: https://github.com/NLnetLabs/domain/pull/425
+[#427]: https://github.com/NLnetLabs/domain/pull/427
+[#440]: https://github.com/NLnetLabs/domain/pull/440
+[#441]: https://github.com/NLnetLabs/domain/pull/441
+[#446]: https://github.com/NLnetLabs/domain/pull/446
+[#459]: https://github.com/NLnetLabs/domain/pull/459
+[#463]: https://github.com/NLnetLabs/domain/pull/463
+[#470]: https://github.com/NLnetLabs/domain/pull/470
+[#472]: https://github.com/NLnetLabs/domain/pull/472
+[#475]: https://github.com/NLnetLabs/domain/pull/475
+[#4775]: https://github.com/NLnetLabs/domain/pull/477
+[@weilence]: https://github.com/weilence
+
+
+## 0.10.4
+
+Released 2025-03-31.
+
+Other changes
+
+* Fix a build issue with [*time*](https://time-rs.github.io/) 0.3.41.
+  ([#505], backported from [#503] by [@PSeitz])
+
+[#503]: https://github.com/NLnetLabs/domain/pull/503
+[#505]: https://github.com/NLnetLabs/domain/pull/505
+[@PSeitz]: https://github.com/PSeitz
 
 
 ## 0.10.3